core(rate): dual-owner rating cycle + favorite/disliked projection — C2 complete
The rating chip goes live end to end, on both owners (playback-adjacent →
parity lockstep, same commit):

- TrackModel favorite/disliked are now resolved from the owner's
  library/signals stores at projection time (CoreView grows the two
  borrows; track_model/queue/search rows all enriched — search via a
  RatingStores pair threaded through search_completed).
- The player fingerprint gains the CURRENT track's rating halves: a rate
  mutation changes no other facet, so the flags themselves re-push
  (pinned by a new publisher test; other rows refresh with the next
  queue snapshot).
- RemoteCommand::Rate dispatches on both owners with identical guards:
  cycle-only (up/down/clear answer not_supported until a sender exists),
  current-track-only (its sole GUI sender binds the player model's
  current track; anything else is unknown_track). The App rides its
  existing Action::CycleRating reducer path; the daemon mirrors the
  cycle transitions in gui_rate (skipping the App-only session-event
  recorder, matching its OS-media rating precedent) and bumps the
  library invalidation counter.
- Parity: the b0 script cycles a full neutral→like→dislike→neutral
  revolution on the current track plus both guard rejections — projected
  models and reasons agree across owners.
- publish.rs crossed its size cap with the new lanes: the test module
  moved to remote/publish/tests.rs (pure move), keeping the ratchet
  meaningful instead of re-blessing.

Gates: fmt/clippy 0, desktop lib 3069 green (parity 9), arch/size green,
gui 213 green.
